Create core-ui RepoCard
