+++ src/features/search/screens/SearchRepositoryScreen.js
@@ -9,6 +9,7 @@ import {
   StyleSheet,
 } from 'react-native';
 import {Ionicons, Octicons, FontAwesome} from '@expo/vector-icons';
+import RepoCard from '../../../general/core-ui/RepoCard';
 
 import type {NavigationScreenProp} from 'react-navigation';
 
@@ -45,50 +46,10 @@ class SearchRepositoryScreen extends Component<Object> {
         fork: true,
       },
     ];
-
-    const {
-      list,
-      repoTitle,
-      repoIconListWrapper,
-      repoIconWrapper,
-      repoIconText,
-    } = styles;
     return (
       <ScrollView style={{flex: 1, backgroundColor: '#fff'}}>
         {repoList.map((repo, index) => {
-          return (
-            <TouchableOpacity
-              key={index}
-              style={list}
-              onPress={() => this._openRepo(repo)}
-            >
-              <View>
-                <Text style={repoTitle}>{repo.fullName}</Text>
-                {repo.description !== '' ? (
-                  <Text>{repo.description}</Text>
-                ) : null}
-                <View style={repoIconListWrapper}>
-                  <View style={repoIconWrapper}>
-                    <Ionicons name="md-star" size={16} color="grey" />
-                    <Text style={repoIconText}>{repo.starsCount}</Text>
-                  </View>
-                  <View style={repoIconWrapper}>
-                    <Octicons name="repo-forked" size={16} color="grey" />
-                    <Text style={repoIconText}>{repo.forksCount}</Text>
-                  </View>
-                  <View style={repoIconWrapper}>
-                    <FontAwesome name="circle" size={16} color="#f9e03b" />
-                    <Text style={repoIconText}>{repo.language}</Text>
-                  </View>
-                </View>
-              </View>
-              {repo.fork ? (
-                <Octicons name="repo-forked" size={32} color="grey" />
-              ) : (
-                <Octicons name="repo" size={32} color="grey" />
-              )}
-            </TouchableOpacity>
-          );
+          return <RepoCard key={index} repo={repo} onPress={this._openRepo} />;
         })}
       </ScrollView>
     );
@@ -98,35 +59,5 @@ class SearchRepositoryScreen extends Component<Object> {
     this.props.navigation.navigate('RepositoryScreen');
   }
 }
-
-const styles = StyleSheet.create({
-  list: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    borderTopWidth: 1,
-    borderColor: '#d6d4d4',
-    padding: 12,
-  },
-  repoTitle: {
-    fontWeight: 'bold',
-  },
-  repoIconListWrapper: {
-    display: 'flex',
-    flexDirection: 'row',
-    marginTop: 4,
-  },
-  repoIconWrapper: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginRight: 8,
-  },
-  repoIconText: {
-    fontSize: 12,
-    marginLeft: 6,
-  },
-});
 
 export default SearchRepositoryScreen;
